Move update schedule
